requirements.txt
discord.py>=2.0
openai-whisper
sentence-transformers
pydub
sounddevice
numpy
matplotlib
